Windows: increase timeout for extent request retries
The AFS Redirector requests file data extents from the afsd_service.exe.  If
it does not receive the requested extent within 10 seconds it issues another
request for that extent.  Extent processing in the afsd_service is handled
by background daemons that process tasks serially from a work queue.   When
the load on the system is large enough that satisfying the work queue takes
longer than 10 seconds, the redirector would retry the request.  This would
increase the length of the work queue and increase lock contention.

Increasing the timeout period for extent retries to two minutes
significantly reduces the number of retry attempts while maintaining
protection against a lost extent request.  Two minutes is selected because
that is the rx hard dead call timeout.

Change-Id: I8169fbdc2d3456f151359ffe9eaa8eeccb2ceaf1
Reviewed-on: http://gerrit.openafs.org/6237
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Jeffrey Altman <jaltman@secure-endpoints.com>
Tested-by: Jeffrey Altman <jaltman@secure-endpoints.com>