Windows: Bad DV invalidate only when new DV not 0
If the current DV is BAD_VERSION and the new DV is 0, do not send
an invalidation to the redirector.  It only results in wasteful work.
If the current DV is BAD_VERSION the object either:

 1. was never previously known

 2. was recently flushed

 3. the cm_scache_t was recycled

In all cases, the redirector does not have knowledge of the object
since either it didn't exist or a previous invalidation was sent.

Reviewed-on: http://gerrit.openafs.org/6392
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Jeffrey Altman <jaltman@secure-endpoints.com>
Tested-by: Jeffrey Altman <jaltman@secure-endpoints.com>
(cherry picked from commit 7a1c8e57a08dcebce2f3b861a0d7bbabcc5272ba)

Change-Id: Icaaa61adf076af43444ec9884cd81da5ed4fa173
Reviewed-on: http://gerrit.openafs.org/6402
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Jeffrey Altman <jaltman@secure-endpoints.com>
Tested-by: Jeffrey Altman <jaltman@secure-endpoints.com>